#6: refactor paypal constants location
